requirements.txt
# Core AI & LLM
anthropic>=0.25.0
openai>=1.30.0
langchain>=0.2.0
langchain-anthropic>=0.1.15

# Web Framework & API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
httpx>=0.27.0

# Task Scheduling
apscheduler>=3.10.4
celery>=5.3.6
redis>=5.0.4

# Data & Storage
sqlalchemy>=2.0.30
alembic>=1.13.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
aiosqlite>=0.20.0

# Integrations
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.129.0
slack-sdk>=3.27.1
notion-client>=2.2.1
stripe>=9.9.0
airtable-python-wrapper>=0.15.3

# Document Generation
reportlab>=4.2.0
jinja2>=3.1.4
weasyprint>=62.3

# Utilities
python-dotenv>=1.0.1
tenacity>=8.3.0
rich>=13.7.1
typer>=0.12.3
structlog>=24.2.0
arrow>=1.3.0

# Web Scraping
playwright>=1.44.0
beautifulsoup4>=4.12.3
lxml>=5.2.2

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-cov>=5.0.0
respx>=0.21.1
faker>=25.2.0
